+++ tests/idl/programs/client-interactions/src/lib.rs
@@ -35,6 +35,18 @@ pub mod client_interactions {
         ctx.accounts.account.enum_field = enum_arg;
         Ok(())
     }
+
+    pub fn type_alias(
+        ctx: Context<TypeAlias>,
+        type_alias_u8: TypeAliasU8,
+        type_alias_u8_array: TypeAliasU8Array,
+        type_alias_struct: TypeAliasStruct,
+    ) -> Result<()> {
+        ctx.accounts.account.type_alias_u8 = type_alias_u8;
+        ctx.accounts.account.type_alias_u8_array = type_alias_u8_array;
+        ctx.accounts.account.type_alias_struct = type_alias_struct;
+        Ok(())
+    }
 }
 
 #[derive(Accounts)]
@@ -81,7 +93,7 @@ pub struct EnumAccount {
 #[derive(AnchorSerialize, AnchorDeserialize, Clone, Copy, Debug, Eq, PartialEq)]
 pub enum MyEnum {
     Unit,
-    Named { x: u64, y: u64 },
+    Named { point_x: u64, point_y: u64 },
     Unnamed(u8, u8, u16, u16),
     UnnamedStruct(MyStruct),
 }
@@ -92,4 +104,28 @@ pub struct MyStruct {
     pub u16: u16,
     pub u32: u32,
     pub u64: u64,
+}
+
+#[derive(Accounts)]
+pub struct TypeAlias<'info> {
+    #[account(zero)]
+    pub account: Account<'info, TypeAliasAccount>,
+}
+
+#[account]
+pub struct TypeAliasAccount {
+    pub type_alias_u8: TypeAliasU8,
+    pub type_alias_u8_array: TypeAliasU8Array,
+    pub type_alias_struct: TypeAliasStruct,
+}
+
+pub type TypeAliasU8 = u8;
+pub type TypeAliasU8Array = [TypeAliasU8; 8];
+pub type TypeAliasStruct = MyStruct;
+
+/// Generic type aliases should not get included in the IDL
+pub type TypeAliasNotSupported<'a, T> = NotSupported<'a, T>;
+pub struct NotSupported<'a, T> {
+    _t: T,
+    _s: &'a str,
 }