+++ ts/packages/anchor/rollup.config.ts
@@ -33,7 +33,6 @@ export default {
   external: [
     "@coral-xyz/borsh",
     "@solana/web3.js",
-    "base64-js",
     "bn.js",
     "bs58",
     "buffer",